+++ extensions/amp-ad-network-fake-impl/0.1/test/test-amp-ad-network-fake-impl.js
@@ -25,9 +25,19 @@ describes.realWin(
     },
   },
   env => {
+    const title = `<title>Hello, world.</title>`;
+    const styleBoilerplate = `<style amp4ads-boilerplate>body{visibility:hidden}</style>`;
+    const ampExperiment = `<script async custom-element=amp-experiment src=https://cdn.ampproject.org/v0/amp-experiment-0.1.js></script>`;
+    const ampAudio = `<script async custom-element=amp-audio src=https://cdn.ampproject.org/v0/amp-audio-0.1.js></script>`;
+    const noscript = `<noscript><style amp-boilerplate> body{-webkit-animation:none;-moz-animation:none;-ms-animation:none; animation:none}</style></noscript>`;
+    const ampRuntimeStyle = `<style amp-runtime i-amphtml-version=42></style>`;
     const ampRuntimeScript = `<script async src=https://cdn.ampproject.org/amp4ads-v0.js></script>`;
+    const fontLink = `<link href=https://fonts.googleapis.com/css?foobar rel=stylesheet type=text/css>`;
+    const crossorigin = `<link crossorigin href=https://fonts.gstatic.com/ rel="dns-prefetch preconnect">`;
     const metaCharset = `<meta charset=utf-8></meta>`;
     const metaViewport = `<meta name=viewport content="width=device-width,minimum-scale=1,initial-scale=1"></meta>`;
+    const ampCustomStyle = `<style amp-custom></style>`;
+    const linkIcon = `<link href=https://example.com/favicon.ico rel=icon>`;
     let doc;
     let win;
     let fakeImplElem;
@@ -55,128 +65,57 @@ describes.realWin(
       expect(fakeImpl.isValidElement()).to.be.true;
     });
 
-    it('generates extensions metadata', () => {
-      const docString = `<html ⚡><head>
-      <script custom-element="amp-font"
-          src="https://cdn.ampproject.org/v0/amp-font-0.1.js"></script>
-      <script custom-element="amp-list"
-          src="https://cdn.ampproject.org/v0/amp-list-latest.js"></script>
-      <script custom-element="amp-list"
-          src="https://cdn.ampproject.org/v0/amp-list-0.1.js"></script>
-      </head><body></body></html>`;
-      fakeImplElem.setAttribute('id', 'i-amphtml-demo-test-3');
-      const reordered = new AmpAdNetworkFakeImpl(
-        fakeImplElem
-      ).transformCreative_(docString);
-      expect(reordered).to.include(
-        `"extensions":[{"custom-element":"amp-font","src":"https://cdn.ampproject.org/v0/amp-font-0.1.js"},{"custom-element":"amp-list","src":"https://cdn.ampproject.org/v0/amp-list-latest.js"}]`
-      );
-    });
-
-    it('generates runtime offsets', () => {
-      const docString =
-        '<!doctype html><html amp4ads><head>' +
+    it('reorders head and inserts metadata into creative', () => {
+      const input =
+        `<html><head>` +
+        title +
+        styleBoilerplate +
+        ampExperiment +
+        ampAudio +
+        noscript +
+        ampRuntimeStyle +
+        ampRuntimeScript +
+        fontLink +
+        crossorigin +
         metaCharset +
         metaViewport +
-        ampRuntimeScript +
-        `</head><body>hello world
-      </script></body></html>`;
-      fakeImplElem.setAttribute('id', 'i-amphtml-demo-test-4');
-      const reordered = new AmpAdNetworkFakeImpl(
-        fakeImplElem
-      ).transformCreative_(docString);
-      expect(reordered).to.include(`"ampRuntimeUtf16CharOffsets":[128,201]`);
-      expect(reordered.substr(128, 73)).to.equal(
-        `<script async="" src="https://cdn.ampproject.org/amp4ads-v0.js"></script>`
-      );
-    });
-
-    it('generates json offsets', () => {
-      const docString =
-        '<!doctype html><html amp4ads><head>' +
-        metaCharset +
-        metaViewport +
-        ampRuntimeScript +
-        '<style amp-custom></style></head><body><amp-analytics><script type=application/json>Iñtërnâtiônàlizætiøn☃💩</script></amp-analytics></body></html>';
-      fakeImplElem.setAttribute('id', 'i-amphtml-demo-test-5');
-      const reordered = new AmpAdNetworkFakeImpl(
-        fakeImplElem
-      ).transformCreative_(docString);
-      expect(reordered).to.include(
-        `"jsonUtf16CharOffsets":{"amp-analytics":[258,322]`
-      );
-      expect(reordered.substr(258, 64)).to.equal(
-        `<script type="application/json">Iñtërnâtiônàlizætiøn☃💩</script>`
-      );
-    });
-
-    it('handles already existing amp-ad-metadata json', () => {
-      const docString =
-        '<!doctype html><html amp4ads><head>' +
-        metaCharset +
-        metaViewport +
-        ampRuntimeScript +
-        '<style amp-custom></style></head><body><script type=application/json amp-ad-metadata>{"diagnosis": "pre-existing condition"}</script></body></html>';
-      fakeImplElem.setAttribute('id', 'i-amphtml-demo-test-5');
-      const reordered = new AmpAdNetworkFakeImpl(
-        fakeImplElem
-      ).transformCreative_(docString);
-      expect(reordered).to.include(`"diagnosis":"pre-existing condition"`);
-    });
-
-    it('generates image metadata', () => {
-      const docString =
-        '<!doctype html><html amp4ads><head>' +
-        metaCharset +
-        metaViewport +
-        ampRuntimeScript +
-        `<style amp-custom></style></head><body>
-             <amp-img notSrc="https://goo.com"></amp-img>
-            <amp-img-other src="https://goo.com"></amp-img>
-            <Amp-imG src="https://some.image.com?a=b" height=999 width=0></amp-img>
-            <div>
-              <amp-img src="https://foo.image.com?a=b&c=d" height=1 width=1></amp-img>
-            </div>
-            <amp-img src="https://test.image.com" height=1 width=5></amp-img>
-            <amp-img src="https://foo.image.com?a=b&c=d" height=1 width=2></amp-img>
-            <aMp-imG src="https://test.image.com?d=e" height=1 width=3></amp-img>
-            <amp-img src="https://test.image.com?f=g" height=4 width=1></amp-img>
-            <amp-img src="https://test.image.com?h=i" height=0 width=1999></amp-img>
-             </body></html>`;
-      fakeImplElem.setAttribute('id', 'i-amphtml-demo-test-5');
-      const reordered = new AmpAdNetworkFakeImpl(
-        fakeImplElem
-      ).transformCreative_(docString);
-      expect(reordered).to.include(
-        `"images":["https://some.image.com?a=b","https://foo.image.com?a=b&c=d","https://test.image.com","https://foo.image.com?a=b&c=d","https://test.image.com?d=e","https://test.image.com?f=g","https://test.image.com?h=i"]`
-      );
-    });
-
-    it('adds the correct metadata for a story ad', () => {
-      const storyAd = `
-    <html ⚡4ads>
-    <head>
-      <meta charset="utf-8">
-      <meta name="viewport" content="width=device-width,minimum-scale=1">
-      <style amp4ads-boilerplate>body{visibility:hidden}</style>
-      <script async src="https://cdn.ampproject.org/amp4ads-v0.js"></script>
-      <meta name="amp-cta-type" content="INSTALL">
-      <meta name="amp-cta-url" content="https://www.amp.dev">
-    </head>
-    <body>Hello, AMP4ADS world.</body>
-    </html>
-    `;
+        ampCustomStyle +
+        linkIcon +
+        `</head><body><amp-analytics><script type=application/json>Iñtërnâtiônàlizætiøn☃💩</script></amp-analytics></body></html>`;
       fakeImplElem.setAttribute('id', 'i-amphtml-demo-test');
-      fakeImplElem.setAttribute('amp-story', '');
       const transformed = new AmpAdNetworkFakeImpl(
         fakeImplElem
-      ).transformCreative_(storyAd);
-      const root = new DOMParser().parseFromString(transformed, 'text/html')
-        .documentElement;
-      const metadata = root.querySelector('[amp-ad-metadata]').textContent;
-      const parsed = JSON.parse(metadata);
-      expect(parsed.ctaType).to.equal('INSTALL');
-      expect(parsed.ctaUrl).to.equal('https://www.amp.dev');
+      ).transformCreative_(input);
+      expect(transformed).to.include(
+        '<script type="application/json" amp-ad-metadata>'
+      );
+      const root = new DOMParser().parseFromString(transformed, 'text/html');
+      const parsed = JSON.parse(
+        root.querySelector('script[amp-ad-metadata]').textContent
+      );
+      const runtimeOffsetStart = parsed.ampRuntimeUtf16CharOffsets[0];
+      const runtimeOffsetEnd = parsed.ampRuntimeUtf16CharOffsets[1];
+      expect(
+        transformed.substr(
+          runtimeOffsetStart,
+          runtimeOffsetEnd - runtimeOffsetStart
+        )
+      ).to.equal(
+        '<script async="" src="https://cdn.ampproject.org/amp4ads-v0.js"></script><script async="" custom-element="amp-experiment" src="https://cdn.ampproject.org/v0/amp-experiment-0.1.js"></script><script async="" custom-element="amp-audio" src="https://cdn.ampproject.org/v0/amp-audio-0.1.js"></script>'
+      );
+      const jsonOffsetStart = parsed.jsonUtf16CharOffsets['amp-analytics'][0];
+      const jsonOffsetEnd = parsed.jsonUtf16CharOffsets['amp-analytics'][1];
+      expect(
+        transformed.substr(jsonOffsetStart, jsonOffsetEnd - jsonOffsetStart)
+      ).to.equal(
+        '<script type="application/json">Iñtërnâtiônàlizætiøn☃💩</script>'
+      );
+      expect(transformed).to.include(
+        '"customElementExtensions":["amp-experiment","amp-audio"]'
+      );
+      expect(transformed).to.include(
+        '"customStyleSheets":[{"href":"https://fonts.googleapis.com/css?foobar"}]'
+      );
     });
   }
 );
